refactor(ProfitCalculator): domain 에서 util로 refactor
